+++ packages/cryptography/test/unit/PublicKey.js
@@ -1,8 +1,3 @@
-import {
-    AccountId,
-    TransactionId,
-    TransferTransaction,
-} from "../../../../src/exports.js";
 import PrivateKey from "../../src/PrivateKey.js";
 import PublicKey from "../../src/PublicKey.js";
 import * as hex from "../../src/encoding/hex.js";
@@ -102,27 +97,5 @@ describe("PublicKey", function () {
         const signature = key.sign(message);
 
         expect(publicKey.verify(message, signature)).to.be.true;
-    });
-
-    it("ECDSA verify transaction", function () {
-        const transaction = new TransferTransaction()
-            .setTransactionId(TransactionId.generate(new AccountId(0, 0, 4)))
-            .freeze();
-
-        const key = PrivateKey.generateECDSA();
-        key.signTransaction(transaction);
-
-        expect(key.getPublicKey().verifyTransaction(transaction)).isTrue();
-    });
-
-    it("ED25519 verify transaction", function () {
-        const transaction = new TransferTransaction()
-            .setTransactionId(TransactionId.generate(new AccountId(0, 0, 4)))
-            .freeze();
-
-        const key = PrivateKey.generateED25519();
-        key.signTransaction(transaction);
-
-        expect(key.getPublicKey().verifyTransaction(transaction)).isTrue();
     });
 });
